+++ client/route/dashboard/reducers/index.js
@@ -1,15 +1,17 @@
 const initState = {
-  time: new Date().toString(),
+  time: new Date().toLocaleTimeString(),
 }
 
 function dashboard(state = initState, action) {
 
   switch(action.type) {
+
     case 'TIME':
-    console.log(1)
-    return Object.assign({}, state, {time: new Date().toString()})
+    return Object.assign({}, state, {time: new Date().toLocaleTimeString()})
+
     default:
     return state
+
   }
 
 }