+++ tasks/publish.js
@@ -9,11 +9,11 @@ function publish() {
       src: '**/*',
       dotfiles: true,
       //add: true,  // never remove existing files
-      branch: 'production',
+      //branch: 'production',
       //message: 'Updates',
       //repo: 'https://example.com/other/repo.git',
       //user: {name: '', email: ''},
-      clone: 'temp',
+      //clone: 'temp',
       //push: true,
       //silent: true,
       //logger: function(string){},